[bug-fix] int->short
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@893 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/phot/phot_optec.c
@@ -45,7 +45,7 @@ phot_command (char command, short arg)
 {
   char cmd_buf[3];
   cmd_buf[0] = command;
-  *((int *) (&(cmd_buf[1]))) = arg;
+  *((short *) (&(cmd_buf[1]))) = arg;
   write (fd, cmd_buf, 3);
 }
